+++ doc/changelog.rst
@@ -20,6 +20,11 @@ Localization
 Admin
 ~~~~~
 
+- Add mass actions to products list
+- Add mass actions to orders list
+- Add mass actions to contacts list
+- Picotable lists now support mass actions.
+
 Addons
 ~~~~~~